refactor: refactor utils
