+++ tagoapi/models/__init__.py
@@ -3,5 +3,6 @@ from .Vehicle import Vehicle
 from .Station import Station
 from .ArrivalInfo import ArrivalInfo
 from .BaseModel import BaseModel
+from .BaseList import BaseList
 
-__all__ = ['Route', 'Vehicle', 'Station', 'ArrivalInfo', 'BaseModel']
+__all__ = ['Route', 'Vehicle', 'Station', 'ArrivalInfo', 'BaseModel', 'BaseList']
