+++ tagoapi/utils/__init__.py
@@ -1,23 +1,19 @@
-from .cache import Cache
-from .cache_util import convert_model
-from .cache_util import cache
-from .parser import parse_metadata
 from .convertor import convert
-from .parser import KeyExtract
+
+from .params import KeyExtract
 from .params import build_params
+
 from .get import http_get
+
 from .get_station import get_station
 
 
 
 __all__ = [
-    "Cache",
-    "KeyExtract",
-    "convert_model",
-    "parse_metadata",
     "convert",
     "build_params",
     "cache",
     "http_get",
-    "get_station"
+    "get_station",
+    "KeyExtract"
 ]
