+++ tagoapi/utils/cache/__init__.py
@@ -1,0 +1,4 @@
+from .cache import cache
+from .cache import _generate_cache_key
+
+__all__ = ["cache", "_generate_cache_key"]
