+++ tagoapi/utils/decorator/__init__.py
@@ -1,0 +1,3 @@
+from convert_model import convert_model
+
+__all__=["convert_model"]
